requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mutagen==1.47.0
numpy==2.1.2
psycopg2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodomex==3.21.0
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
urllib3==2.2.3
websockets==13.1
Werkzeug==3.0.4
WTForms==3.1.2